+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - `hm_arr` -- `hm_arr_shrink_to_fit()`
     - `hm_heap` -- `hm_heap_shrink_to_fit()`
     - `hm_queue` -- `hm_queue_shrink_to_fit()`
+    - `hm_stack` -- `hm_stack_shrink_to_fit()`
 
 ## [0.14.0] - 2026-08-29